+++ src/lib.rs
@@ -106,6 +106,38 @@ pub fn llvm_ir(
 }
 
 ///
+/// Runs the zkEVM assembly mode.
+///
+pub fn zkasm(
+    input_files: &[PathBuf],
+    include_metadata_hash: bool,
+    debug_config: Option<compiler_llvm_context::DebugConfig>,
+) -> anyhow::Result<Build> {
+    let path = match input_files.len() {
+        1 => input_files.first().expect("Always exists"),
+        0 => anyhow::bail!("The input file is missing"),
+        length => anyhow::bail!(
+            "Only one input file is allowed in the zkEVM assembly mode, but found {}",
+            length,
+        ),
+    };
+
+    let project = Project::try_from_zkasm_path(path)?;
+
+    let optimizer_settings = compiler_llvm_context::OptimizerSettings::none();
+    let target_machine = compiler_llvm_context::TargetMachine::new(&optimizer_settings)?;
+    let build = project.compile_all(
+        target_machine,
+        optimizer_settings,
+        false,
+        include_metadata_hash,
+        debug_config,
+    )?;
+
+    Ok(build)
+}
+
+///
 /// Runs the standard output mode.
 ///
 #[allow(clippy::too_many_arguments)]
@@ -218,7 +250,9 @@ pub fn standard_json(
         compiler_llvm_context::OptimizerSettings::try_from(&solc_input.settings.optimizer)?;
 
     let include_metadata_hash = match solc_input.settings.metadata {
-        Some(ref metadata) => metadata.bytecode_hash != compiler_llvm_context::MetadataHash::None,
+        Some(ref metadata) => {
+            metadata.bytecode_hash != Some(compiler_llvm_context::MetadataHash::None)
+        }
         None => true,
     };