+++ src/project/contract/ir/mod.rs
@@ -5,6 +5,7 @@
 pub mod evmla;
 pub mod llvm_ir;
 pub mod yul;
+pub mod zkasm;
 
 use crate::evmla::assembly::Assembly;
 use crate::yul::parser::statement::object::Object;
@@ -12,6 +13,7 @@ use crate::yul::parser::statement::object::Object;
 use self::evmla::EVMLA;
 use self::llvm_ir::LLVMIR;
 use self::yul::Yul;
+use self::zkasm::ZKASM;
 
 ///
 /// The contract source code.
@@ -26,6 +28,8 @@ pub enum IR {
     EVMLA(EVMLA),
     /// The LLVM IR source code.
     LLVMIR(LLVMIR),
+    /// The zkEVM assembly source code.
+    ZKASM(ZKASM),
 }
 
 impl IR {
@@ -49,6 +53,13 @@ impl IR {
     pub fn new_llvm_ir(path: String, source: String) -> Self {
         Self::LLVMIR(LLVMIR::new(path, source))
     }
+
+    ///
+    /// A shortcut constructor.
+    ///
+    pub fn new_zkasm(path: String, source: String) -> Self {
+        Self::ZKASM(ZKASM::new(path, source))
+    }
 }
 
 impl<D> compiler_llvm_context::WriteLLVM<D> for IR
@@ -60,6 +71,7 @@ where
             Self::Yul(inner) => inner.declare(context),
             Self::EVMLA(inner) => inner.declare(context),
             Self::LLVMIR(_inner) => Ok(()),
+            Self::ZKASM(_inner) => Ok(()),
         }
     }
 
@@ -68,6 +80,7 @@ where
             Self::Yul(inner) => inner.into_llvm(context),
             Self::EVMLA(inner) => inner.into_llvm(context),
             Self::LLVMIR(_inner) => Ok(()),
+            Self::ZKASM(_inner) => Ok(()),
         }
     }
 }