+++ src/solc/mod.rs
@@ -40,7 +40,7 @@ impl Compiler {
     pub const FIRST_CLI_VIA_IR_VERSION: semver::Version = semver::Version::new(0, 8, 13);
 
     /// The last supported version of `solc`.
-    pub const LAST_SUPPORTED_VERSION: semver::Version = semver::Version::new(0, 8, 19);
+    pub const LAST_SUPPORTED_VERSION: semver::Version = semver::Version::new(0, 8, 20);
 
     ///
     /// A shortcut constructor.