+++ src/solc/standard_json/input/settings/metadata.rs
@@ -8,11 +8,12 @@ use serde::Serialize;
 ///
 /// The `solc --standard-json` input settings metadata.
 ///
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Default, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Metadata {
     /// The bytecode hash mode.
-    pub bytecode_hash: compiler_llvm_context::MetadataHash,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub bytecode_hash: Option<compiler_llvm_context::MetadataHash>,
 }
 
 impl Metadata {
@@ -20,6 +21,8 @@ impl Metadata {
     /// A shortcut constructor.
     ///
     pub fn new(bytecode_hash: compiler_llvm_context::MetadataHash) -> Self {
-        Self { bytecode_hash }
+        Self {
+            bytecode_hash: Some(bytecode_hash),
+        }
     }
 }